+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/listbox/ListComboBox.java
@@ -18,54 +18,64 @@
  */
 package is.codion.swing.common.ui.component.listbox;
 
-import is.codion.common.reactive.value.AbstractValue;
-import is.codion.common.reactive.value.Value;
-import is.codion.common.reactive.value.ValueSet;
 import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
+import is.codion.swing.common.ui.component.builder.AbstractComponentValueBuilder;
+import is.codion.swing.common.ui.component.value.AbstractComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.common.ui.key.KeyEvents;
 
+import org.jspecify.annotations.Nullable;
+
+import javax.swing.ComboBoxEditor;
+import javax.swing.DefaultListCellRenderer;
 import javax.swing.JComboBox;
 import javax.swing.JComponent;
+import javax.swing.JList;
+import javax.swing.JTextField;
+import javax.swing.ListCellRenderer;
 import javax.swing.event.ListDataEvent;
 import javax.swing.event.ListDataListener;
+
+import java.awt.Component;
+import java.awt.event.ActionListener;
+import java.util.Collections;
 import java.util.LinkedHashSet;
-import java.util.Objects;
 import java.util.Set;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
+import java.util.function.Function;
 
 import static is.codion.swing.common.ui.control.Control.command;
 import static is.codion.swing.common.ui.key.KeyEvents.MENU_SHORTCUT_MASK;
 import static java.awt.event.KeyEvent.VK_DELETE;
 import static java.awt.event.KeyEvent.VK_INSERT;
 import static java.util.Collections.emptySet;
-import static java.util.stream.Collectors.toCollection;
+import static java.util.Objects.requireNonNull;
+import static javax.swing.SwingConstants.CENTER;
 
 final class ListComboBox<T> extends JComboBox<T> {
 
-	private final Value<T> itemValue;
-	private final ValueSet<T> linkedValue;
+	private final ComponentValue<? extends JComponent, T> componentValue;
 
-	ListComboBox(FilterComboBoxModel<T> comboBoxModel, ComponentValue<? extends JComponent, T> itemValue,
-							 ValueSet<T> linkedValue) {
-		super(comboBoxModel);
-		this.itemValue = itemValue;
-		this.linkedValue = linkedValue;
-		new ListBoxItemValue<>(itemValue, comboBoxModel).link(linkedValue);
+	private ListComboBox(DefaultBuilder<T> builder) {
+		super(FilterComboBoxModel.builder()
+						.items(Collections.<T>emptyList())
+						.build());
+		this.componentValue = builder.componentValue;
+		setEditor(new Editor<>(componentValue));
+		setEditable(true);
+		setRenderer(new Renderer<>(horizontalAlignment(componentValue.component()), builder.formatter));
 		KeyEvents.builder()
 						.keyCode(VK_INSERT)
 						.action(command(this::addItem))
-						.enable(itemValue.component());
+						.enable(componentValue.component());
 		KeyEvents.builder()
 						.keyCode(VK_DELETE)
 						.action(command(this::removeItem))
-						.enable(itemValue.component());
+						.enable(componentValue.component());
 		KeyEvents.builder()
 						.keyCode(VK_DELETE)
 						.modifiers(MENU_SHORTCUT_MASK)
 						.action(command(this::clear))
-						.enable(itemValue.component());
+						.enable(componentValue.component());
 	}
 
 	@Override
@@ -73,19 +83,15 @@ final class ListComboBox<T> extends JComboBox<T> {
 		return (FilterComboBoxModel<T>) super.getModel();
 	}
 
-	Value<T> itemValue() {
-		return itemValue;
-	}
-
-	ValueSet<T> linkedValue() {
-		return linkedValue;
+	ComponentValue<? extends JComponent, T> componentValue() {
+		return componentValue;
 	}
 
 	private void addItem() {
 		FilterComboBoxModel<T> comboBoxModel = getModel();
-		if (!itemValue.isNull() && !comboBoxModel.items().contains(itemValue.getOrThrow())) {
-			comboBoxModel.items().add(itemValue.getOrThrow());
-			itemValue.clear();
+		if (!componentValue.isNull() && !comboBoxModel.items().contains(componentValue.getOrThrow())) {
+			comboBoxModel.items().add(componentValue.getOrThrow());
+			componentValue.clear();
 			if (isPopupVisible()) {
 				hidePopup();
 				showPopup();
@@ -107,48 +113,152 @@ final class ListComboBox<T> extends JComboBox<T> {
 		getModel().items().clear();
 	}
 
-	private static final class ListBoxItemValue<T> extends AbstractValue<Set<T>> {
+	private static int horizontalAlignment(JComponent component) {
+		if (component instanceof JTextField) {
+			return ((JTextField) component).getHorizontalAlignment();
+		}
 
-		private final Value<T> itemValue;
-		private final FilterComboBoxModel<T> comboBoxModel;
+		return CENTER;
+	}
 
-		private ListBoxItemValue(Value<T> itemValue, FilterComboBoxModel<T> comboBoxModel) {
-			super(emptySet());
+	private static final class Editor<T> implements ComboBoxEditor {
+
+		private final ComponentValue<? extends JComponent, T> itemValue;
+
+		private Editor(ComponentValue<? extends JComponent, T> itemValue) {
 			this.itemValue = itemValue;
-			this.comboBoxModel = comboBoxModel;
-			this.comboBoxModel.addListDataListener(new ComboBoxModelListener());
-			itemValue.addListener(this::notifyObserver);
 		}
 
 		@Override
-		protected Set<T> getValue() {
-			return Stream.concat(Stream.of(itemValue.get()), IntStream.range(0, comboBoxModel.getSize())
-											.mapToObj(comboBoxModel::getElementAt))
-							.filter(Objects::nonNull)
-							.collect(toCollection(LinkedHashSet::new));
+		public void setItem(Object anObject) {
+			itemValue.set((T) anObject);
 		}
 
 		@Override
-		protected void setValue(Set<T> value) {
-			comboBoxModel.items().clear();
-			value.forEach(comboBoxModel.items()::add);
+		public Component getEditorComponent() {
+			return itemValue.component();
 		}
 
-		private final class ComboBoxModelListener implements ListDataListener {
+		@Override
+		public @Nullable Object getItem() {
+			return itemValue.get();
+		}
 
-			@Override
-			public void intervalAdded(ListDataEvent e) {
-				notifyObserver();
+		@Override
+		public void selectAll() {}
+
+		@Override
+		public void addActionListener(ActionListener l) {}
+
+		@Override
+		public void removeActionListener(ActionListener l) {}
+	}
+
+	private static final class Renderer<T> implements ListCellRenderer<T> {
+
+		private final Function<T, String> formatter;
+
+		private final DefaultListCellRenderer listCellRenderer = new DefaultListCellRenderer();
+
+		private Renderer(int horizontalAlignment, Function<T, String> formatter) {
+			this.formatter = formatter;
+			listCellRenderer.setHorizontalAlignment(horizontalAlignment);
+		}
+
+		@Override
+		public Component getListCellRendererComponent(JList<? extends T> list, T value,
+		                                              int index, boolean isSelected, boolean cellHasFocus) {
+			return listCellRenderer.getListCellRendererComponent(list, formatter.apply(value), index, isSelected, cellHasFocus);
+		}
+	}
+
+	static final class DefaultBuilder<T>
+ 					extends AbstractComponentValueBuilder<JComboBox<T>, Set<T>, ListComboBoxBuilder<T>>
+					implements ListComboBoxBuilder<T> {
+
+		static final ComponentStep ITEM = new DefaultComponentStep();
+
+		private final ComponentValue<? extends JComponent, T> componentValue;
+
+		private Function<T, String> formatter = new DefaultFormatter<>();
+
+		private DefaultBuilder(ComponentValue<? extends JComponent, T> componentValue) {
+			this.componentValue = componentValue;
+		}
+
+		@Override
+		public ListComboBoxBuilder<T> formatter(Function<T, String> formatter) {
+			this.formatter = requireNonNull(formatter);
+			return this;
+		}
+
+		@Override
+		protected JComboBox<T> createComponent() {
+			return new ListComboBox<>(this);
+		}
+
+		@Override
+		protected ComponentValue<JComboBox<T>, Set<T>> createValue(JComboBox<T> comboBox) {
+			return new ListBoxComponentValue<>((ListComboBox<T>) comboBox);
+		}
+
+		private static final class ListBoxComponentValue<T>
+						extends AbstractComponentValue<JComboBox<T>, Set<T>>
+						implements ComponentValue<JComboBox<T>, Set<T>> {
+
+			public ListBoxComponentValue(ListComboBox<T> comboBox) {
+				super(comboBox, emptySet());
+				comboBox.componentValue().addListener(this::notifyObserver);
+				comboBox.getModel().addListDataListener(new ComboBoxModelListener());
 			}
 
 			@Override
-			public void intervalRemoved(ListDataEvent e) {
-				notifyObserver();
+			protected Set<T> getComponentValue() {
+				ListComboBox<T> comboBox = (ListComboBox<T>) component();
+				Set<T> result = new LinkedHashSet<>();
+				comboBox.componentValue().optional().ifPresent(result::add);
+				result.addAll(comboBox.getModel().items().included().get());
+
+				return result;
 			}
 
 			@Override
-			public void contentsChanged(ListDataEvent e) {
-				notifyObserver();
+			protected void setComponentValue(Set<T> value) {
+				((ListComboBox<T>) component()).getModel().items().set(value);
+			}
+
+			private final class ComboBoxModelListener implements ListDataListener {
+
+				@Override
+				public void intervalAdded(ListDataEvent e) {
+					notifyObserver();
+				}
+
+				@Override
+				public void intervalRemoved(ListDataEvent e) {
+					notifyObserver();
+				}
+
+				@Override
+				public void contentsChanged(ListDataEvent e) {
+					notifyObserver();
+				}
+			}
+		}
+
+		private static final class DefaultFormatter<T> implements Function<T, String> {
+
+			@Override
+			public String apply(Object value) {
+				return value == null ? "" : value.toString();
+			}
+		}
+
+		private static final class DefaultComponentStep implements ComponentStep {
+
+			@Override
+			public <T> ListComboBoxBuilder<T> component(ComponentValue<? extends JComponent, T> component) {
+				return new DefaultBuilder<>(requireNonNull(component));
 			}
 		}
 	}